+++ utils/generate_chain_caches.sh
@@ -1,3 +1,2 @@
 set -e -x
-cargo nextest run generate_zcashd_chain_cache --run-ignored ignored-only
 cargo nextest run generate_zebrad_large_chain_cache --run-ignored ignored-only